feat: Add userRelations to schema
This commit adds the `userRelations` function to the `schema.ts` file. The `userRelations` function defines the relationship between the `users` table and the `profiles` table using the `one` method from the `drizzle-orm` library. This change enhances the schema definition and allows for easier querying of related data in the backend.